+++ packages/class-core-test/src/gen/LionCore_M3.g.ts
@@ -467,6 +467,12 @@ export abstract class Classifier extends LanguageEntity {
     replaceFeaturesAtIndex(movedChild: Feature, newIndex: number) {
         this._features.replaceAtIndex(movedChild, newIndex);
     }
+    moveFeaturesOffsetBased(oldIndex: number, indexOffset: number) {
+        this._features.moveOffsetBased(oldIndex, indexOffset);
+    }
+    moveAndReplaceFeaturesOffsetBased(oldIndex: number, indexOffset: number) {
+        this._features.moveAndReplaceOffsetBased(oldIndex, indexOffset);
+    }
 
     protected constructor(classifier: $lwCore.Classifier, id: $lwJson.LionWebId, receiveDelta?: $lwClassCore.DeltaReceiver, parentInfo?: $lwClassCore.Parentage) {
         super(classifier, id, receiveDelta, parentInfo);
@@ -675,6 +681,12 @@ export class Enumeration extends DataType {
     replaceLiteralsAtIndex(movedChild: EnumerationLiteral, newIndex: number) {
         this._literals.replaceAtIndex(movedChild, newIndex);
     }
+    moveLiteralsOffsetBased(oldIndex: number, indexOffset: number) {
+        this._literals.moveOffsetBased(oldIndex, indexOffset);
+    }
+    moveAndReplaceLiteralsOffsetBased(oldIndex: number, indexOffset: number) {
+        this._literals.moveAndReplaceOffsetBased(oldIndex, indexOffset);
+    }
 
     public constructor(classifier: $lwCore.Classifier, id: $lwJson.LionWebId, receiveDelta?: $lwClassCore.DeltaReceiver, parentInfo?: $lwClassCore.Parentage) {
         super(classifier, id, receiveDelta, parentInfo);
@@ -756,6 +768,12 @@ export class Language extends $lwClassCore.NodeBase implements IKeyed {
     }
     replaceEntitiesAtIndex(movedChild: LanguageEntity, newIndex: number) {
         this._entities.replaceAtIndex(movedChild, newIndex);
+    }
+    moveEntitiesOffsetBased(oldIndex: number, indexOffset: number) {
+        this._entities.moveOffsetBased(oldIndex, indexOffset);
+    }
+    moveAndReplaceEntitiesOffsetBased(oldIndex: number, indexOffset: number) {
+        this._entities.moveAndReplaceOffsetBased(oldIndex, indexOffset);
     }
 
     private readonly _dependsOn: $lwClassCore.OptionalMultiReferenceValueManager<Language>;