+++ packages/ts-utils/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.10.0 — not yet released
+## 0.10.0
 
 * Implement a small library for lexicographically comparisons (of strings, mainly): `Comparer` type, `{regular|locale}StringComparer` constants, and `mappedComparer`, `lexiComparer`, and `mappedLexiComparer` functions.
 * Expose a `Sorter` type, and `sorterWith` function for sorting.